Upgrade ESLint to v10 and enable consistent TypeScript type-aware linting (#476)
* Cleaned eth towards new eslint

* Update package.json

* Cleaned scw_js

* Modern eslint

* Further cleaning of the other sub-repos

* fixing plenty of void errors

* Finish the cleaning.

* make everything prettier

* Make it cleaner

* Cleaner package-lock.json

* Update x402_verify.ts

* Clean the workflows

* Cleaner CI

+++ website/components/FacilitatorApproval.tsx
@@ -287,9 +287,9 @@ export function FacilitatorApproval({
     );
   }
 
-  const formattedAllowance = allowance !== undefined ? formatUnits(allowance as bigint, usdcConfig.decimals) : "—";
+  const formattedAllowance = allowance !== undefined ? formatUnits(allowance, usdcConfig.decimals) : "—";
 
-  const hasAllowance = allowance !== undefined && (allowance as bigint) > 0n;
+  const hasAllowance = allowance !== undefined && allowance > 0n;
 
   return (
     <div className={container}>
